+++ scripts/utils/shared.py
